pyproject.toml
[project]
name = "component-custom-python-example-repo-1"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "keboola-component>=1.6.10",
]
